XD-1008 merge OStoreTransactionTest and OStoreTransactionTest1
